Inserted LRU cache, added static configurations
Added the LRU Linked List data structure which automatically checks to
see if the current allocation amount is always below the max.

Added some static methods and variables so that the programmer can
configure how much the max should be (1/4, 1/5, 1/6, 1/7, 1/8) of the
RAM along with which Cache Algorithm to use.

Next step is to write a lot of Log.d messages in the entire library with
the heading if statement "if developer mode equals true" surrounding it.
Easier debugging as the project gets bigger.