LRU and FIFO working correctly
Created a proper test run for both LRU and FIFO. They are working
correctly and seem to produce the correct results in taking up a certain
amount of memory. FIFO has extremely low overhead while LRU is slightly
more overhead. The LRU algorithm goes over the limit just slightly but
its not the end of the world at the moment.

MRU and LFU aren't written to work yet.