Removed LFU, lowered initial overhead, version 1.0 basically ready
This version is basically 1.0. The calculations seems correct and all of
the algorithms behave appropriately. Ended up removing LFU since it did
not function correctly. Will add it later. Since I removed the redundant
AsyncTask and made it all synchronous, it requires significantly less
code. The only thing to do now is setup what an app would like like in
using this.