Added error handling and no need for null handling
Added some error handling to alert the programmer of specific issues.
Most of the stacktraces are turned off but the `Log.e()` messages give
the programmer a good idea of which function threw the error and why.
Researched more about how to handle `null` values in calculating their
respective bit allocation and found out that, at least according to my
research, null values are actually pointers to a null instance which
doesn't appear to take up memory.